Make Aliasable only a requirement of Shareable
The Aliasable constraint counts the nested references.
When `share`ing a resource `a` (s.t. `Shareable a`), we need to make
sure the counter of all nested references of `a` are also incremented.

This may sound a bit contrived, but the point of `Shareable` is to
facilitate sharing larger structures (bigger than `Alias).

So, if we are to make `[Alias a]` `Shareable`, the following should
work:

```
let some_alias :: Alias ()
let l = [some_alias]
case share l of
    (x1:xs1, x2:xs2) -> do
        forget x1
        forget x2
        forget xs1
        forget xs2
```

I think the point is that the implementation of `Shareable` should be
careful about this... which means I don't need `Aliasable` at all, just
a proper implementation of `Shareable`. At the moment, `Aliasable` is
used just to facilitate these implementations.

So.. get rid of Aliasable all together.